Extract shared utilities, add useMember hook, and standardize navigation
Reduce duplication by extracting repeated logic into reusable modules:
- dateUtils.js: 7 date functions (parseUTCDate, formatPhotoDate, formatDateRange, etc.)
- memberUtils.js: 6 member functions (formatName, getInitials, getMilestoneInfo, etc.)
- useMember hook: centralized authenticated member fetching with enabled/deps options
- Replace onNavigate prop-drilling with direct useNavigate() calls across all components
