+++ src/components/MenuButton.jsx
@@ -5,38 +5,15 @@ import MenuItem from '@mui/material/MenuItem'
 import Icon from '@mui/material/Icon'
 import { useNavigate } from 'react-router-dom'
 
-export default function MenuButton({ onNavigate }) {
+export default function MenuButton() {
   const [anchorEl, setAnchorEl] = React.useState(null)
   const open = Boolean(anchorEl)
   const handleClick = (event) => setAnchorEl(event.currentTarget)
   const handleClose = () => setAnchorEl(null)
   const navigate = useNavigate()
 
-  const viewMap = {
-    home: '/',
-    about: '/about',
-    announcements: '/announcements',
-    events: '/events',
-    membership: '/membership',
-    newsletters: '/newsletters',
-    faqs: '/faqs',
-    contact: '/contact',
-    sightings: '/sightings',
-    hotspots: '/hotspots',
-    news: '/news',
-    newsfeed: '/newsfeed',
-    resources: '/resources',
-    officers: '/officers',
-    photos: '/photos'
-  }
-
-  const handleSelect = (value) => {
+  const handleSelect = (path) => {
     handleClose()
-    if (onNavigate) {
-      onNavigate(value)
-      return
-    }
-    const path = viewMap[value] || '/'
     navigate(path)
   }
 
@@ -53,20 +30,20 @@ export default function MenuButton({ onNavigate }) {
         FBC
       </Button>
       <Menu id="app-menu" anchorEl={anchorEl} open={open} onClose={handleClose}>
-        <MenuItem onClick={() => handleSelect('home')}>Home</MenuItem>
-        <MenuItem onClick={() => handleSelect('about')}>About</MenuItem>
-        <MenuItem onClick={() => handleSelect('announcements')}>Club News</MenuItem>
-        <MenuItem onClick={() => handleSelect('events')}>Events</MenuItem>
-        <MenuItem onClick={() => handleSelect('newsletters')}>Newsletters</MenuItem>
-        <MenuItem onClick={() => handleSelect('membership')}>Membership</MenuItem>
-        <MenuItem onClick={() => handleSelect('officers')}>Officers</MenuItem>
-        <MenuItem onClick={() => handleSelect('faqs')}>FAQ's</MenuItem>
-        <MenuItem onClick={() => handleSelect('contact')}>Contact</MenuItem>
-        <MenuItem onClick={() => handleSelect('sightings')}>Sightings</MenuItem>
-        <MenuItem onClick={() => handleSelect('hotspots')}>Hotspots</MenuItem>
-        <MenuItem onClick={() => handleSelect('newsfeed')}>Birding News</MenuItem>
-        <MenuItem onClick={() => handleSelect('resources')}>Resources</MenuItem>
-        <MenuItem onClick={() => handleSelect('photos')}>Photos</MenuItem>
+        <MenuItem onClick={() => handleSelect('/')}>Home</MenuItem>
+        <MenuItem onClick={() => handleSelect('/about')}>About</MenuItem>
+        <MenuItem onClick={() => handleSelect('/announcements')}>Club News</MenuItem>
+        <MenuItem onClick={() => handleSelect('/events')}>Events</MenuItem>
+        <MenuItem onClick={() => handleSelect('/newsletters')}>Newsletters</MenuItem>
+        <MenuItem onClick={() => handleSelect('/membership')}>Membership</MenuItem>
+        <MenuItem onClick={() => handleSelect('/officers')}>Officers</MenuItem>
+        <MenuItem onClick={() => handleSelect('/faqs')}>FAQ's</MenuItem>
+        <MenuItem onClick={() => handleSelect('/contact')}>Contact</MenuItem>
+        <MenuItem onClick={() => handleSelect('/sightings')}>Sightings</MenuItem>
+        <MenuItem onClick={() => handleSelect('/hotspots')}>Hotspots</MenuItem>
+        <MenuItem onClick={() => handleSelect('/newsfeed')}>Birding News</MenuItem>
+        <MenuItem onClick={() => handleSelect('/resources')}>Resources</MenuItem>
+        <MenuItem onClick={() => handleSelect('/photos')}>Photos</MenuItem>
       </Menu>
     </div>
   )
